Fix Bug #490; the status dialog no longer shows negative numbers when the number of calls > 2,147,483,648.
The fix required two changes.  First, make eval_calls a 64-bit integer instead of a 32-bit integer.  Second, change the code which displays eval_calls in the status dialog to properly render 64-bit numbers.

git-svn-id: https://svn.code.sf.net/p/fmslogo/source@4354 9d0050d7-00b0-40f9-871c-8e02cbd5794f

+++ logo/trunk/src/eval.cpp
@@ -85,7 +85,7 @@ FIXNUM g_ValueStatus;
 
 FIXNUM dont_fix_ift = 0;
 
-long eval_count = 0;             // current count of "evaluations" calls
+long long eval_count = 0;        // current count of "evaluations" calls
 
 int g_CatchErrorCount = 0;       // the number of nested blocks of CATCH "ERROR
                                  // This is used to disable "ERRACT processing.
